+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next Version
 
+#### Added
+
+- Scheme Templates [#672](https://github.com/yonaskolb/XcodeGen/pull/672) @bclymer
+
 #### Fixed
 - Fixed macOS unit test setting preset [#665](https://github.com/yonaskolb/XcodeGen/pull/665) @yonaskolb
 - Add `rcproject` files to sources build phase instead of resources [#669](https://github.com/yonaskolb/XcodeGen/pull/669) @Qusic